Fix the removal of export of ids starting with `_`
This bug was introduced just before the commit 5f4f3e6. In practice, the
committed code didn't remove the exports of the ids starting with `_`.
There was no side-effects just had no effect.
Also added `_vars` to the list of exclusions of this process since
it's still being used by chart tasks.